Simplify wrapper by using PERL5LIB
I remove one level of indirection, so that a shell script directly calls
the "perl_binary" script. Before this, I used to have a perl script
being called before, which would mess with @INC.

The idea was to have a place where we could have more control over the
execution of the binaries. For example, we could remove entries from
@INC.

This doesn't seem all that useful at this point though, so just setting
PERL5LIB in the bash script for now seems more than enough.